third_party/blink/web_tests/media/controls/small-overflow-menu-when-pip-disabled.html

<!DOCTYPE html>
<title>Small overflow menu when pip disabled</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-controls.js"></script>
<body>
<video controls></video>
<script>
async_test(t => {
  let video = document.querySelector("video");
  video.src = "../content/test.ogv";
  enableTestMode(video);

  const overflowMenuList = overflowMenu(video);
  const textTrackList = textTrackMenu(video);

  internals.settings.setPictureInPictureEnabled(false);
  assert_false(document.pictureInPictureEnabled, 'pip should be disabled');

  video.onloadedmetadata = t.step_func_done(() => {
    assert_false(overflowMenuList.classList.contains('pip-presented'), 'overflow menu should not have pip-presented');
    assert_false(textTrackList.classList.contains('pip-presented'), 'track list should not have pip-presented');
  });
});
</script>
</body>
